Move class into lib directory
