+++ lib/destroy_unused_providers.rb
